Fix MAD devices not showing up on map correctly
The X and Y coords are swapped currently for MAD devices. This is because X is mapped to latitude and Y is mapped to longitude which is incorrect in this case. 
Some more info: https://gdal.org/tutorials/osr_api_tut.html#crs-and-axis-order
This is a very confusing topic overall when working with geometries and both versions exsists (latitude, longitude / longitude, latitude) depending on the libraries used and also db engines.